+++ i18nextify.js
@@ -89,11 +89,10 @@
   }
 
   function _setPrototypeOf(o, p) {
-    _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) {
+    _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) {
       o.__proto__ = p;
       return o;
     };
-
     return _setPrototypeOf(o, p);
   }
 
@@ -126,7 +125,7 @@
   }
 
   function _getPrototypeOf(o) {
-    _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) {
+    _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) {
       return o.__proto__ || Object.getPrototypeOf(o);
     };
     return _getPrototypeOf(o);
@@ -308,6 +307,13 @@
         return new Logger(this.logger, _objectSpread(_objectSpread({}, {
           prefix: "".concat(this.prefix, ":").concat(moduleName, ":")
         }), this.options));
+      }
+    }, {
+      key: "clone",
+      value: function clone(options) {
+        options = options || this.options;
+        options.prefix = options.prefix || this.prefix;
+        return new Logger(this.logger, options);
       }
     }]);
 
@@ -629,6 +635,7 @@
         }
 
         if (mix === undefined) return undefined;
+        if (mix === null) return null;
 
         if (path.endsWith(p)) {
           if (typeof mix === 'string') return mix;
@@ -1276,7 +1283,7 @@
               } else {
                 var pluralSuffix;
                 if (needsPluralHandling) pluralSuffix = _this4.pluralResolver.getSuffix(code, options.count, options);
-                var zeroSuffix = '_zero';
+                var zeroSuffix = "".concat(_this4.options.pluralSeparator, "zero");
 
                 if (needsPluralHandling) {
                   finalKeys.push(key + pluralSuffix);
@@ -2026,7 +2033,12 @@
           var optionsString = "{".concat(c[1]);
           key = c[0];
           optionsString = this.interpolate(optionsString, clonedOptions);
-          optionsString = optionsString.replace(/'/g, '"');
+          var matchedSingleQuotes = optionsString.match(/'/g);
+          var matchedDoubleQuotes = optionsString.match(/"/g);
+
+          if (matchedSingleQuotes && matchedSingleQuotes.length % 2 === 0 && !matchedDoubleQuotes || matchedDoubleQuotes.length % 2 !== 0) {
+            optionsString = optionsString.replace(/'/g, '"');
+          }
 
           try {
             clonedOptions = JSON.parse(optionsString);
@@ -2157,6 +2169,21 @@
     };
   }
 
+  function createCachedFormatter(fn) {
+    var cache = {};
+    return function invokeFormatter(val, lng, options) {
+      var key = lng + JSON.stringify(options);
+      var formatter = cache[key];
+
+      if (!formatter) {
+        formatter = fn(lng, options);
+        cache[key] = formatter;
+      }
+
+      return formatter(val);
+    };
+  }
+
   var Formatter = function () {
     function Formatter() {
       var options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
@@ -2166,23 +2193,38 @@
       this.logger = baseLogger.create('formatter');
       this.options = options;
       this.formats = {
-        number: function number(val, lng, options) {
-          return new Intl.NumberFormat(lng, options).format(val);
-        },
-        currency: function currency(val, lng, options) {
-          return new Intl.NumberFormat(lng, _objectSpread$4(_objectSpread$4({}, options), {}, {
+        number: createCachedFormatter(function (lng, options) {
+          var formatter = new Intl.NumberFormat(lng, options);
+          return function (val) {
+            return formatter.format(val);
+          };
+        }),
+        currency: createCachedFormatter(function (lng, options) {
+          var formatter = new Intl.NumberFormat(lng, _objectSpread$4(_objectSpread$4({}, options), {}, {
             style: 'currency'
-          })).format(val);
-        },
-        datetime: function datetime(val, lng, options) {
-          return new Intl.DateTimeFormat(lng, _objectSpread$4({}, options)).format(val);
-        },
-        relativetime: function relativetime(val, lng, options) {
-          return new Intl.RelativeTimeFormat(lng, _objectSpread$4({}, options)).format(val, options.range || 'day');
-        },
-        list: function list(val, lng, options) {
-          return new Intl.ListFormat(lng, _objectSpread$4({}, options)).format(val);
-        }
+          }));
+          return function (val) {
+            return formatter.format(val);
+          };
+        }),
+        datetime: createCachedFormatter(function (lng, options) {
+          var formatter = new Intl.DateTimeFormat(lng, _objectSpread$4({}, options));
+          return function (val) {
+            return formatter.format(val);
+          };
+        }),
+        relativetime: createCachedFormatter(function (lng, options) {
+          var formatter = new Intl.RelativeTimeFormat(lng, _objectSpread$4({}, options));
+          return function (val) {
+            return formatter.format(val, options.range || 'day');
+          };
+        }),
+        list: createCachedFormatter(function (lng, options) {
+          var formatter = new Intl.ListFormat(lng, _objectSpread$4({}, options));
+          return function (val) {
+            return formatter.format(val);
+          };
+        })
       };
       this.init(options);
     }
@@ -2200,6 +2242,11 @@
       key: "add",
       value: function add(name, fc) {
         this.formats[name.toLowerCase().trim()] = fc;
+      }
+    }, {
+      key: "addCached",
+      value: function addCached(name, fc) {
+        this.formats[name.toLowerCase().trim()] = createCachedFormatter(fc);
       }
     }, {
       key: "format",
@@ -2342,6 +2389,8 @@
       _this.waitingReads = [];
       _this.maxParallelReads = options.maxParallelReads || 10;
       _this.readingCalls = 0;
+      _this.maxRetries = options.maxRetries >= 0 ? options.maxRetries : 5;
+      _this.retryTimeout = options.retryTimeout >= 1 ? options.retryTimeout : 350;
       _this.state = {};
       _this.queue = [];
 
@@ -2448,7 +2497,7 @@
         var _this3 = this;
 
         var tried = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 0;
-        var wait = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 350;
+        var wait = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : this.retryTimeout;
         var callback = arguments.length > 5 ? arguments[5] : undefined;
         if (!lng.length) return callback(null, {});
 
@@ -2466,19 +2515,19 @@
 
         this.readingCalls++;
         return this.backend[fcName](lng, ns, function (err, data) {
-          if (err && data && tried < 5) {
-            setTimeout(function () {
-              _this3.read.call(_this3, lng, ns, fcName, tried + 1, wait * 2, callback);
-            }, wait);
-            return;
-          }
-
           _this3.readingCalls--;
 
           if (_this3.waitingReads.length > 0) {
             var next = _this3.waitingReads.shift();
 
             _this3.read(next.lng, next.ns, next.fcName, next.tried, next.wait, next.callback);
+          }
+
+          if (err && data && tried < _this3.maxRetries) {
+            setTimeout(function () {
+              _this3.read.call(_this3, lng, ns, fcName, tried + 1, wait * 2, callback);
+            }, wait);
+            return;
           }
 
           callback(err, data);
@@ -2779,7 +2828,7 @@
           options = {};
         }
 
-        if (!options.defaultNS && options.ns) {
+        if (!options.defaultNS && options.defaultNS !== false && options.ns) {
           if (typeof options.ns === 'string') {
             options.defaultNS = options.ns;
           } else if (options.ns.indexOf('translation') < 0) {
@@ -3135,8 +3184,9 @@
           options.lng = options.lng || fixedT.lng;
           options.lngs = options.lngs || fixedT.lngs;
           options.ns = options.ns || fixedT.ns;
+          options.keyPrefix = options.keyPrefix || keyPrefix || fixedT.keyPrefix;
           var keySeparator = _this5.options.keySeparator || '.';
-          var resultKey = keyPrefix ? "".concat(keyPrefix).concat(keySeparator).concat(key) : key;
+          var resultKey = options.keyPrefix ? "".concat(options.keyPrefix).concat(keySeparator).concat(key) : key;
           return _this5.t(resultKey, options);
         };
 
@@ -3272,6 +3322,11 @@
         });
 
         var clone = new I18n(mergedOptions);
+
+        if (options.debug !== undefined || options.prefix !== undefined) {
+          clone.logger = clone.logger.clone(options);
+        }
+
         var membersToCopy = ['store', 'services', 'language'];
         membersToCopy.forEach(function (m) {
           clone[m] = _this8[m];
@@ -3379,6 +3434,8 @@
       fetchApi = global.fetch;
     } else if (typeof window !== 'undefined' && window.fetch) {
       fetchApi = window.fetch;
+    } else {
+      fetchApi = fetch;
     }
   }
 
@@ -3410,6 +3467,8 @@
       fetchApi$1 = global.fetch;
     } else if (typeof window !== 'undefined' && window.fetch) {
       fetchApi$1 = window.fetch;
+    } else {
+      fetchApi$1 = fetch;
     }
   }
 
@@ -3451,18 +3510,8 @@
     return url;
   };
 
-  var requestWithFetch = function requestWithFetch(options, url, payload, callback) {
-    if (options.queryStringParams) {
-      url = addQueryString(url, options.queryStringParams);
-    }
-
-    var headers = defaults({}, typeof options.customHeaders === 'function' ? options.customHeaders() : options.customHeaders);
-    if (payload) headers['Content-Type'] = 'application/json';
-    fetchApi$1(url, defaults({
-      method: payload ? 'POST' : 'GET',
-      body: payload ? options.stringify(payload) : undefined,
-      headers: headers
-    }, typeof options.requestOptions === 'function' ? options.requestOptions(payload) : options.requestOptions)).then(function (response) {
+  var fetchIt = function fetchIt(url, fetchOptions, callback) {
+    fetchApi$1(url, fetchOptions).then(function (response) {
       if (!response.ok) return callback(response.statusText || 'Error', {
         status: response.status
       });
@@ -3473,6 +3522,41 @@
         });
       }).catch(callback);
     }).catch(callback);
+  };
+
+  var omitFetchOptions = false;
+
+  var requestWithFetch = function requestWithFetch(options, url, payload, callback) {
+    if (options.queryStringParams) {
+      url = addQueryString(url, options.queryStringParams);
+    }
+
+    var headers = defaults({}, typeof options.customHeaders === 'function' ? options.customHeaders() : options.customHeaders);
+    if (payload) headers['Content-Type'] = 'application/json';
+    var reqOptions = typeof options.requestOptions === 'function' ? options.requestOptions(payload) : options.requestOptions;
+    var fetchOptions = defaults({
+      method: payload ? 'POST' : 'GET',
+      body: payload ? options.stringify(payload) : undefined,
+      headers: headers
+    }, omitFetchOptions ? {} : reqOptions);
+
+    try {
+      fetchIt(url, fetchOptions, callback);
+    } catch (e) {
+      if (!reqOptions || Object.keys(reqOptions).length === 0 || !e.message || e.message.indexOf('not implemented') < 0) {
+        return callback(e);
+      }
+
+      try {
+        Object.keys(reqOptions).forEach(function (opt) {
+          delete fetchOptions[opt];
+        });
+        fetchIt(url, fetchOptions, callback);
+        omitFetchOptions = true;
+      } catch (err) {
+        callback(err);
+      }
+    }
   };
 
   var requestWithXmlHttpRequest = function requestWithXmlHttpRequest(options, url, payload, callback) {
@@ -3546,6 +3630,8 @@
     if (hasXMLHttpRequest() || typeof ActiveXObject === 'function') {
       return requestWithXmlHttpRequest(options, url, payload, callback);
     }
+
+    callback(new Error('No fetch and no xhr implementation found!'));
   };
 
   function _classCallCheck$1(instance, Constructor) {
@@ -3802,12 +3888,12 @@
     var opt = options || {};
     opt.path = opt.path || '/';
     var value = encodeURIComponent(val);
-    var str = name + '=' + value;
+    var str = "".concat(name, "=").concat(value);
 
     if (opt.maxAge > 0) {
       var maxAge = opt.maxAge - 0;
-      if (isNaN(maxAge)) throw new Error('maxAge should be a Number');
-      str += '; Max-Age=' + Math.floor(maxAge);
+      if (Number.isNaN(maxAge)) throw new Error('maxAge should be a Number');
+      str += "; Max-Age=".concat(Math.floor(maxAge));
     }
 
     if (opt.domain) {
@@ -3815,7 +3901,7 @@
         throw new TypeError('option domain is invalid');
       }
 
-      str += '; Domain=' + opt.domain;
+      str += "; Domain=".concat(opt.domain);
     }
 
     if (opt.path) {
@@ -3823,7 +3909,7 @@
         throw new TypeError('option path is invalid');
       }
 
-      str += '; Path=' + opt.path;
+      str += "; Path=".concat(opt.path);
     }
 
     if (opt.expires) {
@@ -3831,7 +3917,7 @@
         throw new TypeError('option expires is invalid');
       }
 
-      str += '; Expires=' + opt.expires.toUTCString();
+      str += "; Expires=".concat(opt.expires.toUTCString());
     }
 
     if (opt.httpOnly) str += '; HttpOnly';
@@ -3881,7 +3967,7 @@
       document.cookie = serializeCookie(name, encodeURIComponent(value), cookieOptions);
     },
     read: function read(name) {
-      var nameEQ = name + '=';
+      var nameEQ = "".concat(name, "=");
       var ca = document.cookie.split(';');
 
       for (var i = 0; i < ca.length; i++) {
@@ -4084,21 +4170,16 @@
   var subdomain = {
     name: 'subdomain',
     lookup: function lookup(options) {
-      var found;
+      // If given get the subdomain index else 1
+      var lookupFromSubdomainIndex = typeof options.lookupFromSubdomainIndex === 'number' ? options.lookupFromSubdomainIndex + 1 : 1; // get all matches if window.location. is existing
+      // first item of match is the match itself and the second is the first group macht which sould be the first subdomain match
+      // is the hostname no public domain get the or option of localhost
 
-      if (typeof window !== 'undefined') {
-        var language = window.location.href.match(/(?:http[s]*\:\/\/)*(.*?)\.(?=[^\/]*\..{2,5})/gi);
+      var language = typeof window !== 'undefined' && window.location && window.location.hostname && window.location.hostname.match(/^(\w{2,5})\.(([a-z0-9-]{1,63}\.[a-z]{2,6})|localhost)/i); // if there is no match (null) return undefined
 
-        if (language instanceof Array) {
-          if (typeof options.lookupFromSubdomainIndex === 'number') {
-            found = language[options.lookupFromSubdomainIndex].replace('http://', '').replace('https://', '').replace('.', '');
-          } else {
-            found = language[0].replace('http://', '').replace('https://', '').replace('.', '');
-          }
-        }
-      }
+      if (!language) return undefined; // return the given group match
 
-      return found;
+      return language[lookupFromSubdomainIndex];
     }
   };
 
@@ -4111,8 +4192,8 @@
       lookupSessionStorage: 'i18nextLng',
       // cache user language
       caches: ['localStorage'],
-      excludeCacheFor: ['cimode'] //cookieMinutes: 10,
-      //cookieDomain: 'myDomain'
+      excludeCacheFor: ['cimode'] // cookieMinutes: 10,
+      // cookieDomain: 'myDomain'
 
     };
   }
@@ -8124,6 +8205,7 @@
   }
 
   function getDefaults$2() {
+    var scriptEle = document.getElementById('i18nextify');
     return {
       autorun: true,
       ele: document.body,
@@ -8147,7 +8229,8 @@
       namespaceFromPath: false,
       missingKeyHandler: missingHandler,
       ns: [],
-      onInitialTranslate: () => {}
+      onInitialTranslate: () => {},
+      fallbackLng: scriptEle && (scriptEle.getAttribute('fallbacklng') || scriptEle.getAttribute('fallbackLng')) || undefined
     };
   } // auto initialize on dom ready
 
@@ -8240,11 +8323,24 @@
             options.ele.style.display = 'block';
           }
 
+          if (window.document.title) {
+            var keyTitle = window.document.getElementsByTagName('title').length > 0 && window.document.getElementsByTagName('title')[0].getAttribute('i18next-key');
+            window.document.title = instance.t(keyTitle || window.document.title);
+          }
+
+          if (document.querySelector('meta[name="description"]') && document.querySelector('meta[name="description"]').content) {
+            var keyDescr = document.querySelector('meta[name="description"]').getAttribute(instance.options.keyAttr) || document.querySelector('meta[name="description"]').content;
+            document.querySelector('meta[name="description"]').setAttribute("content", instance.t(keyDescr));
+          }
+
           options.onInitialTranslate();
         });
       }
     }
 
+    instance.on('languageChanged', lng => {
+      window.document.documentElement.lang = lng;
+    });
     instance.init(options, done);
 
     if (!domReady) {